pyproject.toml
[tool.poetry]
name = "spark-rvc-inference-module"
version = "0.1.0"
description = "Spark TTS with RVC inference module. A unification of the RVC inference module and Spark TTS projects."
authors = ["vitalii.slobolinskyi"]
license = "MIT"
readme = "README.md"
homepage = "https://github.com/VSlobolinskyi/spark-rvc-inference-module"
repository = "https://github.com/VSlobolinskyi/spark-rvc-inference-module"

[[tool.poetry.packages]]
include = "rvc_ui"
from = "modules"


[tool.poetry.dependencies]
python = ">=3.11,<3.12"

# Common dependencies (non GPU-specific):
fairseq = { git = "https://github.com/One-sixth/fairseq.git" }
joblib = ">=1.1.0"
numba = "*"
llvmlite = "*"
Cython = "*"
numpy = "*"
scipy = "*"
librosa = "==0.10.2"
faiss-cpu = "*"
gradio = "==3.34.0"
soundfile = ">=0.12.1"
ffmpeg-python = ">=0.2.0"
matplotlib = ">=3.7.0"
matplotlib-inline = ">=0.1.3"
praat-parselmouth = ">=0.4.2"
tensorboardX = "*"
tensorboard = "*"
Pillow = ">=9.1.1"
scikit-learn = "*"
tqdm = ">=4.63.1"
uvicorn = ">=0.21.1"
pyworld = "==0.3.2"
onnxruntime = { version = "*", markers = "sys_platform == 'darwin'" }
onnxruntime-gpu = { version = "*", markers = "sys_platform != 'darwin'" }
torchcrepe = "==0.0.23"
fastapi = "==0.88"
torchfcpe = "*"
ffmpy = "==0.3.1"
python-dotenv = ">=1.0.0"
av = "*"

# ---------------------------------------------------------------------------
# --- NVIDIA GPU configuration ---
# torch = [
  # { url = "https://download.pytorch.org/whl/cu118/torch-2.1.1%2Bcu118-cp311-cp311-win_amd64.whl#sha256=d99be44487d3ed0f7e6ef5d6689a37fb4a2f2821a9e7b59e7e04002a876a667a", markers = "sys_platform == 'win32'" },
  # { url = "https://download.pytorch.org/whl/cu118/torch-2.1.1%2Bcu118-cp311-cp311-linux_x86_64.whl#sha256=f3c0ba02b50d0021ff26f030e22d4c45965537cf91f322e52a65b8c58396f81c", markers = "sys_platform == 'linux'" }
# ]
# torchaudio = [
  # { url = "https://download.pytorch.org/whl/cu118/torchaudio-2.1.1%2Bcu118-cp311-cp311-win_amd64.whl#sha256=79b5afa556063be18de4a1964339242301fe04e782e1030a22695257dd9afbd2", markers = "sys_platform == 'win32'" },
  # { url = "https://download.pytorch.org/whl/cu118/torchaudio-2.1.1%2Bcu118-cp311-cp311-linux_x86_64.whl#sha256=2b077639f240176bb27e964e2e9b3a5c2a8d560a3a7bc1ffd0a024e81f2e10b4", markers = "sys_platform == 'linux'" }
# ]
# --- AMD GPU configuration ---
torch = "2.4.1"
torchaudio = "2.4.1"
torch-directml = "^0.2.5.dev240914"
# ---------------------------------------------------------------------------

# Depndenciees for temp_tools
autoflake = "^2.3.1"

[tool.poetry.group.dev.dependencies]
black = "^25.1.0"
pytest = "^7.0"

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
